+++ packages/plugin-admin-api/test/settings-api.spec.js
@@ -85,6 +85,9 @@ describe("settings api", () => {
             },
           },
         },
+        variantHandlers: {
+          register: [],
+        },
       });
     });
   });