+++ test/core-e2e/src/fixtures/config-file-with-routes-handler/mocks.config.js
@@ -3,7 +3,9 @@ const CustomRoutesHandler = require("./CustomRoutesHandler");
 
 module.exports = {
   log: "silly",
-  routesHandlers: [CustomRoutesHandler],
+  variantHandlers: {
+    register: [CustomRoutesHandler],
+  },
   mock: {
     collections: {
       selected: "custom-users",