+++ test/core-e2e/src/route-handler.spec.js
@@ -19,13 +19,13 @@ const {
 
 const RouteHandler = require("./fixtures/config-file-with-routes-handler/CustomRoutesHandler");
 
-describe("when adding route handlers", () => {
+describe("when registering route handlers", () => {
   describe("When started", () => {
     let core;
 
     beforeAll(async () => {
       core = createCore();
-      core.addRoutesHandler(RouteHandler);
+      core.variantHandlers.register([RouteHandler]);
       await startExistingCore(core, fixturesFolder("custom-routes-handler"), {
         mock: { collections: { selected: "custom-users" } },
       });