+++ test/plugin-inquirer-cli-e2e/src/main/fixtures/mocks.config.js
@@ -11,6 +11,10 @@ module.exports = {
     // Allow unknown arguments
     //allowUnknownArguments: false,
   },
+  variantHandlers: {
+    // Variant Handlers to be registered
+    //register: [],
+  },
   plugins: {
     // Plugins to be registered
     //register: [null],