+++ packages/plugin-proxy/CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 ### Removed
 
+## [unreleased]
+
+- refactor: Use new core method `variantHandlers.register` instead of the deprecated `addRoutesHandler`
+
 ## [3.0.0] - 2022-07-05
 
 ### Changed